VimwikiTask: dont remove recur from recur instance
If a task is an instance of a recurring task prototype, it will
currently not be marked as recurring with [R]. Thus its completed_mark
can be ' '. In that case taskwiki tried to set 'recur' to None, which is
equivalent to `task 42 modify recur:` and causes an error:
> You cannot remove the recurrence from a recurring task.

Setting the status of a parent recurring task to 'pending' will be
ignored as well, so the removed line was useless.