Fix: Modifications on task in multiple viewports
Currently, after modifying a task, it's updating only the task line
where the command is issued (like \td for "done") and if the same task
is being shown in another viewport, that line is not reflected with
modifications.
So, on saving the file, it reverts the modifications.

To handle, updating the entire buffer on task modifications