Controverial commit: Delaying setup execution for multi-syntax tests.
This is needed after the fixes on the initialization of vimwiki's
plugin.
It was working before because the sanity_check was restarting the tests
and running a teardown and setup for all these tests (TaskCache was not
being initialized correctly because of vimwiki's load failing).
I do not know of any better way to do this. Maybe there is some way to
pass arguments to the setup method (so we know what configuration
variable to set on taskwiki for the markup_syntax).

I might revisit this in the near future. And see if I can find any
better alternative.