tests: Fix setting vim up twice
Since 8d8fd2c20bdd26d4c3da73e8fd3c18e9f83abfba, the first `setup`
invocation sets g:taskwiki_markup_syntax="None" which is then detected
in `check_sanity` and vim is restarted, this time with correct
g:taskwiki_markup_syntax. This is wasteful and weird, but makes a bit of
sense as a hack to make `MultiSyntaxIntegrationTest` work as the
test_syntax fixture is not available in `setup`.

To fix this:

 * make `teardown` idempotent
 * let `check_sanity` fail early if the client is not set up yet
 * disable `setup` in `MultiSyntaxIntegrationTest`
 * let the sanity check and restart logic handle the setup in
   `MultiSyntaxIntegrationTest` once `markup` is available