tests: Refactor Dockerfile
* Switch from Fedora to Alpine. This makes it possible to have smaller
  docker images (Alpine is smaller by itself; apk add --virtual makes it
  feasible to correctly remove build dependencies) which will be useful
  for caching the images somewhere instead of building them in CI again
  and again. This is expected to cut the CI test time in half.

* Make versions of python, vim and vimwiki configurable via docker build
  args to enable testing of additional environments. This makes the
  build slower, but we will fix that by caching.

Unfortunate consequence of doing all the installs in one RUN step is
that docker build doesn't cache the intermediate steps and therefore
modifying the Dockerfile and testing those modifications locally is now
slower than it used to be. This can likely be improved by using
multi-stage builds (and using BuildKit and its inline cache export to
not break the caching we're about to add a few commits later).