vwtask: Refactor apply_defaults
5bf0c89ae885f41448f321750a09854f3f0a0dd9 left a misleading comment in
the code:

                # If port was detected, break the search

            break

Refactor the code to make it clearer what's going on: we're looking for
the first parent header/preset/viewport of this task, and therefore we
always break. By separating the search and processing of the found
header/preset/viewport, we avoid the confusion.

Related: https://github.com/tools-life/taskwiki/pull/288