syntax: Move setl viewoptions to ftplugin
Doesn't belong in syntax, and shouldn't be set unless we need it for
mkview, so move it next to that mkview.

(BTW, it's a bit confusing that we use setlocal for a global option. Are
we hoping it's going to be made local one day? If not, we may as well
change it to `set` to avoid confusion.)