syntax: Move setl foldtext to BufWinEnter autocmd
Vimwiki now resets fold{method,text} on BufWinEnter:

https://github.com/vimwiki/vimwiki/blob/8d4cb7f11d5f69c90139cd28fa1a3e31e43390e8/plugin/vimwiki.vim#L177
https://github.com/vimwiki/vimwiki/blob/8d4cb7f11d5f69c90139cd28fa1a3e31e43390e8/plugin/vimwiki.vim#L325

so we should do that as well otherwise we get overridden.

If the user actually configured vimwiki to use its folding, by setting
g:vimwiki_folding to something else than '' or 'custom', we should not
replace that. It would be nice if we could somehow augment
VimwikiFoldText and leave the rest intact, but we can't, so it's best to
keep off. Using foldmethod=syntax gets extremely laggy with just a few
hundreds of lines, so we must not force it.

Also, we now need to increase foldlevel in tests because the folding is
no longer being overridden in the default vimwiki configuration, so we'd
get closed folds in tests...